thought controller created

+++ controllers/user-controller.js
@@ -1,4 +1,4 @@
-const { User } = require('../models')
+const { User, Thought } = require('../models')
 
 const userController = {
     getUsers(req, res) {
@@ -13,7 +13,7 @@ const userController = {
         User.findOne({_id: req.params.userId})
         .select('-__v')
         .populate('friends')
-        // .populate('thought')
+        .populate('thought')
         .then((dbUser) => {
             if (!dbUser) {
                 return res.json({message: "User does not exist"})
@@ -43,25 +43,38 @@ const userController = {
         })
     },
 
-    deleteUser(req, res) {
-        User.findOneAndDelete(
-            {_id: req.params.userId},
-        )
-        .then((dbUser) => {
-            if (!dbUser) {
-                return res.json({message: "User does not exist"})
-            }
-            res.json(dbUser)
+    deleteUser({ params }, res) {
+        User.findOneAndDelete({ _id: params.userId })
+        .then((deletedUserData) => {
+          if (!deletedUserData) {
+            res.status(404).json({ message: "There is no user with this ID" });
+            return;
+          }
+  
+          Thought.deleteMany({ username: deletedUserData.username })
+            .then(res.json({ message: "User and associated thoughts deleted successfully!" }))
+            .catch((err) => res.status(400).json(err));
         })
     },
 
-    addFriend(req, res) {
+    addFriend({ params }, res) {
+        User.findOneAndUpdate(
+          { _id: params.userId },
+          { $push: { friends: params.friendId } },
+          { new: true }
+        )
+          .then((data) => res.json(data))
+      },
 
-    
-    },
+      removeFriend({ params }, res) {
+        User.findOneAndUpdate(
+          { _id: params.userId },
+          { $pull: { friends: params.friendId } },
+          { new: true }
+        )
+          .then((data) => res.json(data))
+      },
 
-    removeFriend(req,res) {}
-
-}
+};
 
 module.exports = userController
